- added user input to package install, extract command to mods and other tweaks - changed the event calls from eval() to call_user_func()

+++ trunk/includes/classes/fileop.class.php
@@ -150,36 +150,38 @@ class fileop {
 
 
 	function extract_zip($archive,$path='') {
+		$basename=false;
 		if (substr($archive,-4)=='.zip') {
-			require_once dirname(__FILE__).'/zip.class.php';
-
-			$basename=substr(basename($archive),0,-4);
-			$zipfile=new zipfile();
-			$zipfile->read_zip($archive);
-			if (empty($path)) {
-				$path=dirname($archive);
-			}
-			$path.='/'.$basename;
-			if (!is_dir($path)) {
-				$this->mkdir($path);
-			}
-			for ($i=0;isset($zipfile->dirs[$i]);++$i) {
-				$temp=explode('/',$zipfile->dirs[$i]);
-				$sub_path=$path;
-				for ($j=0;isset($temp[$j]);++$j) {
-					if (!empty($temp[$j])) {
-						$sub_path.='/'.$temp[$j];
-						if (!is_dir($sub_path)) {
-							$this->mkdir($sub_path);
+			if (is_file($archive)) {
+				require_once dirname(__FILE__).'/zip.class.php';
+				
+				$basename=substr(basename($archive),0,-4);
+				$zipfile=new zipfile();
+				$zipfile->read_zip($archive);
+				if (empty($path)) {
+					$path=dirname($archive);
+				}
+				$path.='/'.$basename;
+				if (!is_dir($path)) {
+					$this->mkdir($path);
+				}
+				for ($i=0;isset($zipfile->dirs[$i]);++$i) {
+					$temp=explode('/',$zipfile->dirs[$i]);
+					$sub_path=$path;
+					for ($j=0;isset($temp[$j]);++$j) {
+						if (!empty($temp[$j])) {
+							$sub_path.='/'.$temp[$j];
+							if (!is_dir($sub_path)) {
+								$this->mkdir($sub_path);
+							}
 						}
 					}
 				}
+				for ($i=0;isset($zipfile->files[$i]);++$i) {
+					$this->file_put_contents($path.$zipfile->files[$i]['dir'].'/'.$zipfile->files[$i]['name'],$zipfile->files[$i]['data']);
+				}
 			}
 		}
-		for ($i=0;isset($zipfile->files[$i]);++$i) {
-			$this->file_put_contents($path.$zipfile->files[$i]['dir'].'/'.$zipfile->files[$i]['name'],$zipfile->files[$i]['data']);
-		}
-
 		return $basename;
 	}
